chore (refs T22614): move document Components to client/js.
