+++ client/js/components/document/ElementsAdminItem.vue
