+++ client/js/components/document/ElementsList/FileInfo.vue
