+++ client/js/components/document/ElementsList/IconPublished.vue
